+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ...
 
+### Fixed
+
+- AlterColumnType command now shows as IsImpossible when column is part of a view or table valued function
+
 ## [6.0.0] - 2021-07-28
 
 ### Changed